move tests over, handle lack of organization id
